move legend-data logic to own module
