+++ src/components/legend/index.js
@@ -75,91 +75,6 @@ legend.texts = function(context, td, d, i, traces) {
     else text.call(textLayout);
 };
 
-legend.getLegendData = function(calcdata, opts) {
-
-    // build an { legendgroup: [cd0, cd0], ... } object
-    var lgroupToTraces = {},
-        lgroups = [],
-        hasOneNonBlankGroup = false,
-        slicesShown = {},
-        lgroupi = 0;
-
-    var cd, cd0, trace, lgroup, i, j, labelj;
-
-    function addOneItem(legendGroup, legendItem) {
-        // each '' legend group is treated as a separate group
-        if(legendGroup==='' || !isGrouped(opts)) {
-            var uniqueGroup = '~~i' + lgroupi; // TODO: check this against fullData legendgroups?
-            lgroups.push(uniqueGroup);
-            lgroupToTraces[uniqueGroup] = [[legendItem]];
-            lgroupi++;
-        }
-        else if(lgroups.indexOf(legendGroup) === -1) {
-            lgroups.push(legendGroup);
-            hasOneNonBlankGroup = true;
-            lgroupToTraces[legendGroup] = [[legendItem]];
-        }
-        else lgroupToTraces[legendGroup].push([legendItem]);
-    }
-
-    for(i = 0; i < calcdata.length; i++) {
-        cd = calcdata[i];
-        cd0 = cd[0];
-        trace = cd0.trace;
-        lgroup = trace.legendgroup;
-
-        if(!legendGetsTrace(trace) || !trace.showlegend) continue;
-
-        if(Plots.traceIs(trace, 'pie')) {
-            if(!slicesShown[lgroup]) slicesShown[lgroup] = {};
-            for(j = 0; j < cd.length; j++) {
-                labelj = cd[j].label;
-                if(!slicesShown[lgroup][labelj]) {
-                    addOneItem(lgroup, {
-                        label: labelj,
-                        color: cd[j].color,
-                        i: cd[j].i,
-                        trace: trace
-                    });
-
-                    slicesShown[lgroup][labelj] = true;
-                }
-            }
-        }
-
-        else addOneItem(lgroup, cd0);
-    }
-
-    // won't draw a legend in this case
-    if(!lgroups.length) return [];
-
-    // rearrange lgroupToTraces into a d3-friendly array of arrays
-    var lgroupsLength = lgroups.length,
-        ltraces,
-        legendData;
-
-    if(hasOneNonBlankGroup && isGrouped(opts)) {
-        legendData = new Array(lgroupsLength);
-        for(i = 0; i < lgroupsLength; i++) {
-            ltraces = lgroupToTraces[lgroups[i]];
-            legendData[i] = isReversed(opts) ? ltraces.reverse() : ltraces;
-        }
-    }
-    else {
-        // collapse all groups into one if all groups are blank
-        legendData = [new Array(lgroupsLength)];
-        for(i = 0; i < lgroupsLength; i++) {
-            ltraces = lgroupToTraces[lgroups[i]][0];
-            legendData[0][isReversed(opts) ? lgroupsLength-i-1 : i] = ltraces;
-        }
-        lgroupsLength = 1;
-    }
-
-    // needed in repositionLegend
-    opts._lgroupsLength = lgroupsLength;
-    return legendData;
-};
-
 legend.draw = function(td) {
     var fullLayout = td._fullLayout;
